backend/requirements.txt
aniso8601==9.0.1
asgiref==3.8.1
astroid==2.14.2
async-timeout==4.0.3
attrs==24.2.0
autobahn==24.4.2
Automat==24.8.1
certifi==2022.12.7
cffi==1.17.1
channels==4.1.0
channels-redis==4.2.0
charset-normalizer==3.0.1
constantly==23.10.4
coreapi==2.3.3
coreschema==0.0.4
coverage==7.2.0
cryptography==43.0.1
Cython==0.29.37
daphne==4.1.2
dill==0.3.6
Django==5.1.1
djangorestframework==3.14.0
drf-yasg==1.21.5
graphene==3.2.1
graphene-django==3.0.0
graphql-core==3.2.3
graphql-relay==3.2.0
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.4
incremental==24.7.2
inflection==0.5.1
isort==5.12.0
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
msgpack==1.0.8
openapi-codec==1.3.2
packaging==23.0
platformdirs==3.0.0
promise==2.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pylint==2.16.2
pylint-django==2.5.3
pylint-plugin-utils==0.7
pyOpenSSL==24.2.1
pytz==2022.7.1
PyYAML==5.3.1
redis==5.0.8
requests==2.28.2
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
service-identity==24.1.0
simplejson==3.18.3
six==1.16.0
sqlparse==0.4.3
text-unidecode==1.3
tomli==2.0.1
tomlkit==0.11.6
Twisted==24.7.0
txaio==23.1.1
typing_extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
wrapt==1.14.1
zope.interface==7.0.3
